Solved the amplitudes bug: tapering
 - Reduced the length of the rising part of the windowing function to TWO samples. Does not produce artifacts yet and the kernels look a lot more like kernels.
 - Updated tests to check more things.
 - The actual bug was not found by any unit_test, since all functions were working as they were supposed to.
 - Updated both for plot_wavefields and calc_kernels

I have found the source of the amplitude bug in the kerner code
(Remember: Amplitudes in the central part of the kernels were too high
compared to source and receiver region, while the overall shape looked
okay).

In a nutshell: The problem was the tapering before the FT. By default,
it used a Gaussian window with a flat top of 50% the signal length.
However, that meant reducing the amplitude of the beginning of the
signals of forward and backward field.
Since the beginning of the signal corresponds to the source and the
receiver region, the amplitudes of the kernels were damped there, which
explains the strange shapes, where only the middle part of the kernel
seemed to have normal amplitudes.

The problem can be solved by increasing the length of the flat part of
the tapering window. However, this does not solve it completely. Having
ANY tapering at all will always reduce the kernel amplitude near source
and receiver. This is something we want anyway, but maybe it would be
better to do in a controlled way and not just accidentally.

We cannot go without tapering, since the wavefield time series do not
decay to zero, but how much tapering is the bare minimum?