Submit script for the kerner
Should be the new default way of creating rundirs and jobs for the kerner. The idea is that all values of the old inparam_basic can be set as arguments to submit.py. At the same time, you can also set an input file (in the format of the old inparam_basic) from which submit should take its values. If you do neither, default values are used.

I think this should be quite useful for production runs in the following way:

 - you define all values that do not change in your inparam_basic file (like AxiSEM dirs, convergence criteria)
 - all values that usually change (receiver files etc) are given as arguments to submit.py

This should allow it to quickly start a lot of slightly modified runs without having to worry about creation of directories etc (which is done by submit.py).

The script comes with an extensive help (python ./submit.py -h)

Some keywords in the input file were modified for brevity. Also, I wanted to ensure that all booleans can be set to false by default